added queue

+++ MyaDiscordBot/Commands/Base/CommandHandler.cs
@@ -179,13 +179,17 @@ namespace MyaDiscordBot.Commands.Base
                 using (ILifetimeScope scope = Data.Instance.Container.BeginLifetimeScope())
                 {
                     IAntiSpamService antiSpam = scope.Resolve<IAntiSpamService>();
-                    if (antiSpam.IsSpam(message))
+                    if(message.Author.Id != _client.CurrentUser.Id)
                     {
-                        GuildEmote angry = _client.Guilds.FirstOrDefault(x => x.Id == 783913792668041216).Emotes.Where(x => x.Name.Contains("angry")).Last();
-                        _ = await message.ReplyAsync("請唔好Spam！" + message.Author.Mention + angry.ToString());
-                        await message.DeleteAsync();
-                        return;
+                        if (antiSpam.IsSpam(message))
+                        {
+                            GuildEmote angry = _client.Guilds.FirstOrDefault(x => x.Id == 783913792668041216).Emotes.Where(x => x.Name.Contains("angry")).Last();
+                            _ = await message.ReplyAsync("請唔好Spam！" + message.Author.Mention + angry.ToString());
+                            await message.DeleteAsync();
+                            return;
+                        }
                     }
+
                     if (await antiSpam.IsScam(message))
                     {
                         GuildEmote angry = _client.Guilds.FirstOrDefault(x => x.Id == 783913792668041216).Emotes.Where(x => x.Name.Contains("angry")).Last();
